PreprocessingPipeline

PreprocessingPipeline.Preprocessing.Scaler.BaseScaler — Type

BaseScaler

It is an abstract type for all scalers.

PreprocessingPipeline.Preprocessing.Scaler.StandardScaler — Type

struct StandardScaler

A structure representing a standard scaler.
The standard scaler standardizes features by removing the mean.

# Fields
- `mean::Union{Vector{Float64}, Float64, Nothing}`: Mean of the features.
- `std::Union{Vector{Float64}, Float64, Nothing}`: Standard deviation of the features.

PreprocessingPipeline.Preprocessing.Scaler.MinMaxScaler — Type

struct MinMaxScaler

A structure representing a min-max scaler.
The min-max scaler scales features to a given range.
The range is defined by the `feature_range` parameter.

# Fields
- `feature_range::Tuple{<:Real, <:Real}`: The range to scale the features to.
- `min::Union{Vector{<:Real}, <:Real, Nothing}`: Minimum value of the features.
- `max::Union{Vector{<:Real}, <:Real, Nothing}`: Maximum value of the features.

PreprocessingPipeline.Preprocessing.Scaler.MaxAbsScaler — Type

MaxAbsScaler

Max Abs Scaler is a scaler that scales the data using the maximum absolute value of the data.

# Examples
```julia
scaler = MaxAbsScaler()
```

PreprocessingPipeline.Preprocessing.Normalizers.BaseNormalizer — Type

BaseNormalizer It is an abstract type for all normalizers.

PreprocessingPipeline.Preprocessing.Normalizers.StandardNormalizer — Type

StandardNormalizer

Standard Normalizer is a normalizer that normalizes the data using the standard normalization technique.

# Arguments
- `type::String`: The type of normalization to apply. It can be one of the following:
    - "l1": L1 normalization.
    - "l2": L2 normalization.
    - "max": Max normalization.

# Examples
```julia
normalizer = StandardNormalizer("l2")
```

PreprocessingPipeline.TransformerModule.fit! — Function

fit!(transformer::Transformer, X::Matrix{Any})

Fits the transformer to the given data X. This method should be implemented for any concrete subtype of Transformer.

Arguments

transformer::Transformer: The transformer instance.

X::Matrix{Any}: The input data to fit the transformer.

Returns

The fitted transformer.

fit!(scaler::StandardScaler, X::Vector{<:Real})

Fit the standard scaler to the data.

# Arguments
- `scaler::StandardScaler`: An instance of `StandardScaler`.
- `X::Vector{<:Real}`: The data to fit the scaler.

# Returns
The fitted `StandardScaler`.

fit!(scaler::StandardScaler, X::Matrix{<:Real})

Fit the standard scaler to the data.

# Arguments
- `scaler::StandardScaler`: An instance of `StandardScaler`.
- `X::Matrix{<:Real}`: The data to fit the scaler.

# Returns
The fitted `StandardScaler`.

fit!(scaler::MinMaxScaler, X::Vector{<:Real})

Fit the min-max scaler to the data.

# Arguments
- `scaler::MinMaxScaler`: An instance of `MinMaxScaler`.
- `X::Vector{<:Real}`: The data to fit the scaler.

# Returns
The fitted `MinMaxScaler`.

fit!(scaler::MinMaxScaler, X::Matrix{<:Real})

Fit the min-max scaler to the data.

# Arguments
- `scaler::MinMaxScaler`: An instance of `MinMaxScaler`.
- `X::Matrix{<:Real}`: The data to fit the scaler.

# Returns
The fitted `MinMaxScaler`.

fit!(scaler::MaxAbsScaler, X::Vector{<:Real})

Fit the max-abs scaler to the data.

# Arguments
- `scaler::MaxAbsScaler`: An instance of `MaxAbsScaler`.
- `X::Vector{<:Real}`: The data to fit the scaler.

# Returns
The fitted `MaxAbsScaler`.

fit!(scaler::MaxAbsScaler, X::Matrix{<:Real})

Fit the max-abs scaler to the data.

# Arguments
- `scaler::MaxAbsScaler`: An instance of `MaxAbsScaler`.
- `X::Matrix{<:Real}`: The data to fit the scaler.

# Returns
The fitted `MaxAbsScaler`.

fit!(normalizer::StandardNormalizer, X::Vector{<:Real})

Fit the normalizer to the data.

# Arguments
- `normalizer::StandardNormalizer`: An instance of `StandardNormalizer`.
- `X::Vector{<:Real}`: The data to fit the normalizer.

# Returns
The fitted `StandardNormalizer`.

fit!(normalizer::StandardNormalizer, X::Matrix{<:Real})

Fit the normalizer to the data.

# Arguments
- `normalizer::StandardNormalizer`: An instance of `StandardNormalizer`.
- `X::Matrix{<:Real}`: The data to fit the normalizer.

# Returns
The fitted `StandardNormalizer`.

fit!(encoder::OneHotEncoder, X::Vector{<:String})

Fit the encoder to the data.

# Arguments
- `encoder::OneHotEncoder`: An instance of `OneHotEncoder`.
- `X::Vector{<:String}`: The data to fit the encoder.

# Returns
The fitted `OneHotEncoder`.

fit!(pipeline::Pipeline, X::Matrix{Any})

Fit the pipeline to the input data by fitting each step sequentially. Updates the transformers in pipeline.named_steps with the results of their fit! method.

Arguments

pipeline::Pipeline: The pipeline to fit.

X::Matrix{Any}: Input data matrix.

Returns

The updated pipeline.

PreprocessingPipeline.TransformerModule.transform — Function

transform(transformer::Transformer, X::Matrix{Any})

Applies the transformation defined by the transformer to the input data X. This method should be implemented for any concrete subtype of Transformer.

Arguments

transformer::Transformer: The transformer instance.

X::Matrix{Any}: The input data to transform.

Returns

A transformed version of X.

transform(scaler::StandardScaler, X::Vector{<:Real})

Transform the data using the fitted standard scaler.

# Arguments
- `scaler::StandardScaler`: An instance of `StandardScaler`.
- `X::Vector{<:Real}`: The data to transform.

# Returns
The transformed data.

transform(scaler::StandardScaler, X::Matrix{<:Real})

Transform the data using the fitted standard scaler.

# Arguments
- `scaler::StandardScaler`: An instance of `StandardScaler`.
- `X::Matrix{<:Real}`: The data to transform.

# Returns
The transformed data.

transform(scaler::MinMaxScaler, X::Vector{<:Real})

Transform the data using the min-max scaler.

# Arguments
- `scaler::MinMaxScaler`: An instance of `MinMaxScaler`.
- `X::Vector{<:Real}`: The data to transform.

# Returns
The transformed data.

transform(scaler::MinMaxScaler, X::Matrix{<:Real})

Transform the data using the min-max scaler.

# Arguments
- `scaler::MinMaxScaler`: An instance of `MinMaxScaler`.
- `X::Matrix{<:Real}`: The data to transform.

# Returns
The transformed data.

transform(scaler::MaxAbsScaler, X::Vector{<:Real})

Transform the data using the max-abs scaler.

# Arguments
- `scaler::MaxAbsScaler`: An instance of `MaxAbsScaler`.
- `X::Vector{<:Real}`: The data to transform.

# Returns
The transformed data.

transform(scaler::MaxAbsScaler, X::Matrix{<:Real})

Transform the data using the max-abs scaler.

# Arguments
- `scaler::MaxAbsScaler`: An instance of `MaxAbsScaler`.
- `X::Matrix{<:Real}`: The data to transform.

# Returns
The transformed data.

transform(normalizer::StandardNormalizer, X::Vector{<:Real})

Transform the data using the fitted normalizer.

# Arguments
- `normalizer::StandardNormalizer`: An instance of `StandardNormalizer`.
- `X::Vector{<:Real}`: The data to transform.

# Returns
The transformed data.

transform(normalizer::StandardNormalizer, X::Matrix{<:Real})

Transform the data using the fitted normalizer.

# Arguments
- `normalizer::StandardNormalizer`: An instance of `StandardNormalizer`.
- `X::Matrix{<:Real}`: The data to transform.

# Returns
The transformed data.

transform(encoder::OneHotEncoder, X::Vector{<:String})

Transform the data using the encoder.

# Arguments
- `encoder::OneHotEncoder`: An instance of `OneHotEncoder`.
- `X::Vector{<:String}`: The data to transform.

# Returns
The transformed data.

transform(pipeline::Pipeline, X::Matrix{Any})

Transform the input data using the pipeline by applying each step sequentially.

Arguments

pipeline::Pipeline: The pipeline to use for transformation.

X::Matrix{Any}: Input data matrix.

Returns

Transformed data matrix.

PreprocessingPipeline.TransformerModule.inverse_transform — Function

inverse_transform(scaler::StandardScaler, X::Vector{<:Real})

Inverse transform the data using the fitted standard scaler.

# Arguments
- `scaler::StandardScaler`: An instance of `StandardScaler`.
- `X::Vector{<:Real}`: The data to inverse transform.

# Returns
The inverse transformed data.

inverse_transform(scaler::StandardScaler, X::Matrix{<:Real})

Inverse transform the data using the fitted standard scaler.

# Arguments
- `scaler::StandardScaler`: An instance of `StandardScaler`.
- `X::Matrix{<:Real}`: The data to inverse transform.

# Returns
The inverse transformed data.

inverse_transform(scaler::MinMaxScaler, X::Vector{<:Real})

Inverse transform the data using the min-max scaler.

# Arguments
- `scaler::MinMaxScaler`: An instance of `MinMaxScaler`.
- `X::Vector{<:Real}`: The data to inverse transform.

# Returns
The inverse transformed data.

inverse_transform(scaler::MinMaxScaler, X::Matrix{<:Real})

Inverse transform the data using the min-max scaler.

# Arguments
- `scaler::MinMaxScaler`: An instance of `MinMaxScaler`.
- `X::Matrix{<:Real}`: The data to inverse transform.

# Returns
The inverse transformed data.

inverse_transform(scaler::MaxAbsScaler, X::Vector{<:Real})

Inverse transform the data using the max-abs scaler.

# Arguments
- `scaler::MaxAbsScaler`: An instance of `MaxAbsScaler`.
- `X::Vector{<:Real}`: The data to inverse transform.

# Returns
The inverse transformed data.

inverse_transform(scaler::MaxAbsScaler, X::Matrix{<:Real})

Inverse transform the data using the max-abs scaler.

# Arguments
- `scaler::MaxAbsScaler`: An instance of `MaxAbsScaler`.
- `X::Matrix{<:Real}`: The data to inverse transform.

# Returns
The inverse transformed data.

inverse_transform(normalizer::StandardNormalizer, X::Vector{<:Real})

Inverse transform the data using the fitted normalizer.

# Arguments
- `normalizer::StandardNormalizer`: An instance of `StandardNormalizer`.
- `X::Vector{<:Real}`: The data to inverse transform.

# Returns
The inverse transformed data.

inverse_transform(normalizer::StandardNormalizer, X::Matrix{<:Real})

Inverse transform the data using the fitted normalizer.

# Arguments
- `normalizer::StandardNormalizer`: An instance of `StandardNormalizer`.
- `X::Matrix{<:Real}`: The data to inverse transform.

# Returns
The inverse transformed data.

inverse_transform!(encoder::OneHotEncoder, X::Matrix{<:Real})

Inverse transform the data using the encoder.

# Arguments
- `encoder::OneHotEncoder`: An instance of `OneHotEncoder`.
- `X::Matrix{<:Real}`: The data to inverse transform.

# Returns
The inverse transformed data.